+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-* Changed `database/sql` driver `prepare` behaviour to `nop` with proxing call to conn exec/query with keep-in-cache flag 
+* Changed `database/sql` driver `prepare` behaviour to `nop` with proxing call to conn exec/query with keep-in-cache flag
+* Added metadata to `trace.Driver.OnInvoke` and `trace.Driver.OnNewStream` done events
 
 ## v3.34.0
 * Improved the `xsql` errors mapping to `driver.ErrBadConn`